refactor: 모든 interface를 DTO로 정리
